Learning objectives according to taxonomy of Bloom
Learning objectives can be associated with Courses, Sections or Lessons.
Each one of them must be an object that indicates its relation with the taxonomy of Bloom: to which domain belongs and the concrete learning objective according to that taxonomy. In addition, a field is included for the specific description of the objective.